Implement search on DocBook documents
DocBookPage was a bit of an oddball in that it returned a single Page
object for the entire document, which would create multiple HTML output
files. This is OK for building HTML, but doesn't play well with the
search API. I introduced a DocBookSubPage which provides metadata for
individual pages, while still leaving the entire HTML generation to
the main DocBookPage.