Use lxml.etree.XSLT instead of calling xsltproc
When there are a lot of individual transforms to run, this saves time
by only parsing the stylesheet once. However, when there are only a
handful of transform to run for a given stylesheet, lxml.etree.XSLT
is measurably slower. This is an obvious improvement for Mallard. We
might consider reverting to xsltproc for DocBook in the future.