Separate sources from build directory to allow multiple sources.
I tried using a separate class for merge sources, but the idea of the
directory being the source was baked into a lot of places. In the end,
it was cleaner to create a separate Source class and make everything
aware of that.

Since this broke API, I also took the chance to clean up some API I
was unhappy with. Most notably:

* The is_special_path method on directories has been replaced by the
  create_sources method on Source. Source extensions can decide for
  themselves if they can do something with a config group and create
  Source objects as necessary.

* The get_pages_dir and get_pages methods on Page have been replaced
  by the create_pages method on Page. This works more like get_pages_dir,
  in that it isn't given individual files. But it works on sources, not
  directories. Page extensions can decide for themselves what pages they
  can provide based on the source and the files in it, and create Page
  objects as necessary.

* EmptyDirectory is gone, because it's no longer necessary.

Also, there are docstrings now, so that's a plus.