Added before_script, config_script, and after_script
config_script is renamed from configger. Added a bunch of environment variables
to pass to these scripts. This involved retooling the way Site and Config objects
are initialized, but there shouldn't be any impact outside this repo.

Also added in the --no-translation option to disable translations without fiddling
with the config file. I've been meaning to add this for a while, and it just sort
of fell in here.