Allow Mallard directories to be processed as a single stack
Rather than running an XSLT transform for each page, we create a Mallard
stack that includes all the pages in the directory, and we do a single
XSLT run on that. This is WAY faster. This requires recent changes in
yelp-xsl to de-globalize locale params. And I still need a way to access
Pintail source file params without the global params, mostly likely as
attrs on the cache.